pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ubik-tui"
version = "1.0.0"
description = "UBIK-TUI — Full-screen terminal UI for UBIK-CLI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Damien Ledoux"}]
dependencies = [
    "ubik-cli>=1.0.0",
    "textual>=0.80.0",
]

[project.scripts]
ubik-tui = "ubik_tui.main:main"

[tool.setuptools.packages.find]
where = ["src"]
